@@ -6,12 +6,12 @@ homepage: https://github.com/fluttercommunity/flutter_whatsnew
 maintainer: Rody Davis (@rodydavis)
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_markdown: ^0.3.4
+  flutter_markdown: ^0.6.1
 
 dev_dependencies:
   flutter_test:
